+++ srcs/ddos/header_buffering.c
@@ -10,11 +10,11 @@ extern int g_num_threads;
 
 
 // session counting
-__u64 g_headbuf_num_total;
-__u64 g_headbuf_num_generated_in_sec;
+uint32_t g_headbuf_num_total;
+uint32_t g_headbuf_num_generated_in_sec;
 // from main()
 InputArguments g_headbuf_input;
-__u32 g_headbuf_request_per_sec;
+uint32_t g_headbuf_request_per_sec;
 // for masking next ip address
 MaskingArguments g_headbuf_now;
 // thread
